% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLong.R, R/dataLong_m.R
\docType{data}
\name{dataLong}
\alias{dataLong}
\title{Simulated multiple longitudinal data}
\format{
A data frame which contains id, obstime, x1, x2, w1, w2, and Y1,...,Y10.
\describe{
  \item{id}{patients identifier}
  \item{obstime}{observed times for longitudinal measerements}
  \item{Y1}{the first longitudinal response}
  \item{Y2}{the second longitudinal response}
  \item{Y3}{the third longitudinal response}
  \item{Y4}{the fourth longitudinal response}
  \item{Y5}{the fifth longitudinal response}
  \item{Y6}{the sixth longitudinal response}
  \item{Y7}{the seventh longitudinal response}
  \item{Y8}{the eighth longitudinal response}
  \item{Y9}{the ninth longitudinal response}
  \item{Y10}{the tenth longitudinal response}
  \item{w1}{a continuous covariate}
  \item{x1}{a binary covariate}
  \item{w2}{a continuous covariate}
  \item{x2}{a binary covariate}
}

A data frame which contains id, obstime, x1, x2, w1, w2, and Y1,...,Y10.
\describe{
  \item{id}{patients identifier}
  \item{obstime}{observed times for longitudinal measerements}
  \item{Y1}{the first longitudinal response including missing values}
  \item{Y2}{the second longitudinal response including missing values}
  \item{Y3}{the third longitudinal response}
  \item{Y4}{the fourth longitudinal response}
  \item{Y5}{the fifth longitudinal response}
  \item{Y6}{the sixth longitudinal response}
  \item{Y7}{the seventh longitudinal response}
  \item{Y8}{the eighth longitudinal response}
  \item{Y9}{the ninth longitudinal response}
  \item{Y10}{the tenth longitudinal response}
  \item{w1}{a continuous covariate}
  \item{x1}{a binary covariate}
  \item{w2}{a continuous covariate}
  \item{x2}{a binary covariate}
}
}
\usage{
dataLong

dataLong
}
\description{
Simulated data were generated in which "Y1",...,"Y10" follows a Gaussian distribution, "obstime" represents the observed time with some covariates.

Simulated data were generated in which "Y1",...,"Y10" follows a Gaussian distribution, "obstime" represents the observed time with some covariates.
}
\seealso{
\code{\link{UJM}}

\code{\link{UJM}}
}
\keyword{datasets}
