% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSurv.R
\docType{data}
\name{dataSurv}
\alias{dataSurv}
\title{Simulated competing risks data}
\format{
A data frame which contains id, survtime, CR, w1, x1, w2, x2.
\describe{
  \item{id}{patients identifier}
  \item{survtime}{survival time (the response variable)}
  \item{CR}{event indicator, 1=the first cause, 2=the second cause, 0=censored}
  \item{x1}{a binary covariate}
  \item{x2}{a continuous covariate}
  \item{w1}{a binary covariate}
  \item{w2}{a continuous covariate}
}
}
\usage{
dataSurv
}
\description{
Simulated survival data were generated in which "survtime" follows a survival time with a event indicator called "CR" and some covariates.
}
\seealso{
\code{\link{UJM}}
}
\keyword{datasets}
