% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.R
\name{DP}
\alias{DP}
\title{Dynamic prediction}
\usage{
DP(
  object,
  object2,
  Method = "LBFDR",
  s = s,
  t = t,
  cause_main = cause_main,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  DIC = TRUE,
  quiet = FALSE,
  dataLong,
  dataSurv
)
}
\arguments{
\item{object}{an object inheriting from class VS}

\item{object2}{an object inheriting from class VS2}

\item{Method}{the method for variable selection including "LBFDR" for LBFDR and "BF" for Bayes factor.}

\item{s}{the landmark time for prediction}

\item{t}{the window of prediction for prediction}

\item{cause_main}{the main cause for prediction}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=var(deviance) / 2 is used.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}
}
\value{
\itemize{
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
}
}
\description{
Dynamic prediction for VSJM
}
\details{
Estimate DP for joint modeling based on VS
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
