% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPOM.R
\name{DPOM}
\alias{DPOM}
\title{Dynamic prediction for one marker JM}
\usage{
DPOM(
  object,
  N_marker = 1,
  s = s,
  t = t,
  cause_main = cause_main,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  DIC = TRUE,
  quiet = FALSE,
  dataLong,
  dataSurv
)
}
\arguments{
\item{object}{an object inheriting from class VS}

\item{N_marker}{the number of marker}

\item{s}{the landmark time for prediction}

\item{t}{the window of prediction for prediction}

\item{cause_main}{the main cause for prediction}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=var(deviance) / 2 is used.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}
}
\value{
\itemize{
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 25\% list of posterior median for each parameter
\item 50\% list of posterior median for each parameter
\item 75\% list of posterior median for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
}
}
\description{
Dynamic prediction for One-marker JM
}
\details{
Estimate DP for joint modeling based
}
\examples{
\dontrun{
rm(list = ls())
data(dataLong_ind)
data(dataSurv_ind)
set.seed(2)
INDTRAIN <- sample(dataSurv_ind$id, 0.5 * (dim(dataSurv_ind)[1]))
INDVALID <- dataSurv_ind$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong_ind,
  dataLong_ind$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv_ind,
  dataSurv_ind$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong_ind,
  dataLong_ind$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv_ind,
  dataSurv_ind$id \%in\% INDVALID
)


formFixed <- list(
  Y1 ~ obstime + x1 + x2, Y2 ~ obstime + x1 + x2, Y3 ~ obstime + x1 + x2,
  Y4 ~ obstime + x1 + x2, Y5 ~ obstime + x1 + x2, Y6 ~ obstime + x1 + x2,
  Y7 ~ obstime + x1 + x2, Y8 ~ obstime + x1 + x2, Y9 ~ obstime + x1 + x2,
  Y10 ~ obstime + x1 + x2
)
formRandom <- list(
  ~obstime, ~obstime, ~obstime, ~obstime, ~obstime, ~obstime,
  ~obstime, ~obstime, ~obstime, ~obstime
)
formGroup <- list(~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1 + w2 + x2
model <- list(
  "linear", "linear", "linear", "linear", "linear", "linear",
  "linear", "linear", "quadratic", "quadratic"
)


VS <- VS(formFixed, formRandom, formGroup, formSurv,
  nmark = 2, K1 = 15, K2 = 15,
  model = model, n.chains1 = 2, n.iter1 = 20, n.burnin1 = 10,
  n.thin1 = 1, n.chains2 = 2, n.iter2 = 20, n.burnin2 = 10,
  n.thin2 = 1, simplify = TRUE, Obstime = "obstime", Method = "DS",
  ncl = 2,
  DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

VS$sim_step1[[1]]$PMean$alpha
VS$sim_step1[[2]]$PMean$alpha

SVS(VS)


D1 <- DPOM(VS,
  N_marker = 1, s = 0.1, t = 0.5, cause_main = 1, n.chains = 1,
  n.iter = 10, n.burnin = 5,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE, dataLong_v, dataSurv_v
)



Criteria(
  s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
  CR = dataSurv_v$CR, P = D1$DP$est, cause = 1
)


D2 <- DPOM(VS,
  N_marker = 2, s = 0.1, t = 0.5, cause_main = 1, n.chains = 1,
  n.iter = 1000, n.burnin = 500,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE, dataLong_v, dataSurv_v
)

Criteria(
  s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
  CR = dataSurv_v$CR, P = D2$DP$est, cause = 1
)


rm(list = ls())
data(dataLong)
data(dataSurv)
set.seed(2)
INDTRAIN <- sample(dataSurv$id, 0.9 * (dim(dataSurv)[1]))
INDVALID <- dataSurv$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong,
  dataLong$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv,
  dataSurv$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong,
  dataLong$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv,
  dataSurv$id \%in\% INDVALID
)


formFixed <- list(
  Y1 ~ obstime + x1 + x2, Y2 ~ obstime + x1 + x2, Y3 ~ obstime + x1 + x2,
  Y4 ~ obstime + x1 + x2, Y5 ~ obstime + x1 + x2, Y6 ~ obstime + x1 + x2,
  Y7 ~ obstime + x1 + x2, Y8 ~ obstime + x1 + x2, Y9 ~ obstime + x1 + x2,
  Y10 ~ obstime + x1 + x2
)
formRandom <- list(
  ~obstime, ~obstime, ~obstime, ~obstime, ~obstime, ~obstime,
  ~obstime, ~obstime, ~obstime, ~obstime
)
formGroup <- list(~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1 + w2 + x2
model <- list(
  "linear", "linear", "linear", "linear", "linear", "linear",
  "linear", "linear", "quadratic", "quadratic"
)


VS <- VS(formFixed, formRandom, formGroup, formSurv,
  nmark = 10, K1 = 15, K2 = 15,
  model = model, n.chains1 = 2, n.iter1 = 100, n.burnin1 = 50,
  n.thin1 = 1, n.chains2 = 2, n.iter2 = 60, n.burnin2 = 50,
  n.thin2 = 1, simplify = TRUE, Obstime = "obstime", Method = "DS", ncl = 5,
  DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

VS$sim_step1[[1]]$PMean$alpha
VS$sim_step1[[2]]$PMean$alpha

SVS(VS)


D1 <- DPOM(VS,
  N_marker = 1, s = 0.1, t = 0.5, cause_main = 1, n.chains = 1,
  n.iter = 1000, n.burnin = 500,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE, dataLong_v, dataSurv_v
)

Criteria(
  s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
  CR = dataSurv_v$CR, P = D1$DP$est, cause = 1
)


D2 <- DPOM(VS,
  N_marker = 10, s = 0.1, t = 0.5, cause_main = 1, n.chains = 1,
  n.iter = 1000, n.burnin = 500,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE, dataLong_v, dataSurv_v
)

Criteria(
  s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
  CR = dataSurv_v$CR, P = D2$DP$est, cause = 1
)

}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
