% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVS.R
\name{SVS}
\alias{SVS}
\title{Summary function of VS}
\usage{
SVS(object)
}
\arguments{
\item{object}{an object inheriting from class VS}
}
\value{
\itemize{
\item gamma list of posterior mean for each gamma including estimation, standard deviation, LBFDR and BF
\item alpha list of posterior mean for each alpha including estimation, standard deviation, LBFDR and BF
}
}
\description{
Computing dynamic prediction based on the output of summary of the outputs of the VS function, a two-stage approach to efficiently select variables for joint modeling of multiple longitudinal markers and competing risks outcomes within a Bayesian framework.
}
\details{
It presents a summary of the output of VS function, including LBFDR, BF, and parameter estimations.
}
\examples{
rm(list=ls())
data(dataLong)
data(dataSurv)
set.seed(2)
INDTRAIN <- sample(dataSurv$id, 0.5 * (dim(dataSurv)[1]))
INDVALID <- dataSurv$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong,
  dataLong$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv,
  dataSurv$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong,
  dataLong$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv,
  dataSurv$id \%in\% INDVALID
)



formFixed <- list(Y1 ~ obstime, Y2 ~ obstime, Y3 ~ obstime)
formRandom <- list(~obstime, ~obstime, ~obstime)
formGroup <- list(~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
model <- list("intercept", "linear", "quadratic")



VS <- VS(formFixed, formRandom, formGroup, formSurv,
  nmark = 3, K1 = 15, K2 = 15,
  model = model, n.chains = 1, n.iter = 100, n.burnin = 50,
  n.thin = 1, Obstime = "obstime", Method = "DS", ncl = 2,
  DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

SVS <- SVS(VS)

Step2 <- VS2(VS,
  Method = "LBFDR", n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1, dataLong = dataLong_t, dataSurv = dataSurv_t
)

DP <- DP(VS, Step2,
  Method = "LBFDR", s = 0.1, t = 0.5, n.chains = 1, n.iter = 100, n.burnin = 50,
  n.thin = 1,cause_main=1,
  DIC = TRUE, quiet = FALSE, dataLong = dataLong_v, dataSurv = dataSurv_v
)



######################################################################
\dontrun{
library(DPCri);library(survival)

Criteria(s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
         CR = dataSurv_v$CR, P = DP$DP[,2], cause = 1)

######################################################################
rm(list=ls())
data(dataLong_m)
data(dataSurv_m)
set.seed(2)
INDTRAIN <- sample(dataSurv_m$id, 0.5 * (dim(dataSurv_m)[1]))
INDVALID <- dataSurv_m$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong_m,
  dataLong_m$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv_m,
  dataSurv_m$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong_m,
  dataLong_m$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv_m,
  dataSurv_m$id \%in\% INDVALID
)



formFixed <- list(Y1 ~ obstime, Y2 ~ obstime, Y3 ~ obstime)
formRandom <- list(~obstime, ~obstime, ~obstime)
formGroup <- list(~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
model <- list("intercept", "linear", "quadratic")



VS <- VS(formFixed, formRandom, formGroup, formSurv,
         nmark = 3, K1 = 15, K2 = 15,
         model = model, n.chains = 1, n.iter = 50, n.burnin = 20,
         n.thin = 1, Obstime = "obstime", Method = "DS", ncl = 2,
         DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

SVS <- SVS(VS)

Step2 <- VS2(VS,
             Method = "LBFDR", n.chains = 2, n.iter = 100, n.burnin = 50,
             n.thin = 1, dataLong = dataLong_t, dataSurv = dataSurv_t
)

DP <- DP(VS, Step2,
         Method = "LBFDR", s = 0.1, t = 0.5, n.chains = 1, n.iter = 100, n.burnin = 50,
         n.thin = 1,cause_main=1,
         DIC = TRUE, quiet = FALSE, dataLong = dataLong_v, dataSurv = dataSurv_v
)


Criteria(s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
         CR = dataSurv_v$CR, P = DP$DP[,2], cause = 1)

######################################################################

  rm(list=ls())
  set.seed(2)
  INDTRAIN=sample(dataSurv$id,.9*(dim(dataSurv)[1]))
  dataLong_t=subset(dataLong,
                    dataLong$id \%in\% INDTRAIN)

  dataSurv_t=subset(dataSurv,
                    dataSurv$id \%in\% INDTRAIN)
  names(dataSurv_t)
  formFixed <- list(
    Y1 ~ obstime + x1 + x2, Y2 ~ obstime + x1 + x2, Y3 ~ obstime + x1 + x2,
    Y4 ~ obstime + x1 + x2, Y5 ~ obstime + x1 + x2, Y6 ~ obstime + x1 + x2,
    Y7 ~ obstime + x1 + x2, Y8 ~ obstime + x1 + x2, Y9 ~ obstime + x1 + x2,
    Y10 ~ obstime + x1 + x2
  )
  formRandom <- list(
    ~obstime, ~obstime, ~obstime, ~obstime, ~obstime, ~obstime,
    ~obstime, ~obstime, ~obstime, ~obstime
  )
  formGroup <- list(~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id)
  formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
  model <- list(
    "linear", "linear", "linear", "linear", "linear", "linear",
    "linear", "linear", "quadratic", "quadratic"
  )



  VS_DS <- VS(formFixed, formRandom, formGroup, formSurv,
              nmark = 10, K1 = 15, K2 = 15,
              model = model, n.chains = 1, n.iter = 2000, n.burnin = 1000,
              n.thin = 1, Obstime = "obstime", Method = "DS",ncl=4,
              DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
  )

  SVS_DS <- SVS(VS_DS)



  VS_CS <- VS(formFixed, formRandom, formGroup, formSurv,
              nmark = 10, K1 = 15, K2 = 15,
              model = model, n.chains = 1, n.iter = 2000, n.burnin = 1000,
              n.thin = 1, Obstime = "obstime", Method = "CS",
              DIC = TRUE, quiet = FALSE, dataLong, dataSurv
  )

  SVS_CS <- SVS(VS_CS)

}




}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
