% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UJM.R
\name{UJM}
\alias{UJM}
\title{One-marker Joint Modeling}
\usage{
UJM(
  formFixed,
  formRandom,
  formGroup,
  formSurv,
  dataLong,
  dataSurv,
  K = K,
  model = "linear",
  Obstime = "Obstime",
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  DIC = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{formFixed}{formula for fixed part of longitudinal model}

\item{formRandom}{formula for random part of longitudinal model}

\item{formGroup}{formula specifying the cluster variable for Y (e.g. = ~ subject)}

\item{formSurv}{formula for survival model}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}

\item{K}{Number of nodes and weights for calculating Gaussian quadrature.}

\item{model}{the model for the longitudinal part which includes "intercept", "linear" or "quadratic".}

\item{Obstime}{the observed time in longitudinal data}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=var(deviance) / 2 is used.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}
}
\value{
\itemize{
\item MCMC chains for the unknown parameters
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 25\% list of posterior median for each parameter
\item 50\% list of posterior median for each parameter
\item 75\% list of posterior median for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
}
}
\description{
Fits Bayesian linear or quadratic one-marker joint modeling of longitudinal measurements and competing risks using MCMC
}
\details{
Use the 'JAGS' software to estimate Bayesian linear or quadratic one-marker joint modeling of longitudinal measurements and competing risks using MCMC
}
\examples{
library(survival)
data(dataLong)
data(dataSurv)

\dontrun{

A=UJM(
  formFixed = Y1 ~ obstime + x1 + x2, formRandom = ~obstime,
  formGroup = ~id, formSurv = Surv(survtime, CR) ~ w1 + x1,
  dataLong, dataSurv, K = 15, model = "intercept", Obstime = "obstime",
  n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE
)

UJM(
  formFixed = Y1 ~ obstime + x1 + x2, formRandom = ~obstime,
  formGroup = ~id, formSurv = Surv(survtime, CR) ~ w1 + x1,
  dataLong, dataSurv, K = 15, model = "linear", Obstime = "obstime",
  n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE
)


UJM(
  formFixed = Y1 ~ obstime + x1 + x2, formRandom = ~obstime,
  formGroup = ~id, formSurv = Surv(survtime, CR) ~ w1 + x1,
  dataLong, dataSurv, K = 15, model = "quadratic", Obstime = "obstime",
  n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE
)


UJM(
  formFixed = Y1 ~ obstime, formRandom = ~obstime,
  formGroup = ~id, formSurv = Surv(survtime, CR) ~ w1 + x1,
  dataLong, dataSurv, K = 15, model = "intercept", Obstime = "obstime",
  n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE
)

UJM(
  formFixed = Y1 ~ obstime, formRandom = ~obstime,
  formGroup = ~id, formSurv = Surv(survtime, CR) ~ w1 + x1,
  dataLong, dataSurv, K = 15, model = "linear", Obstime = "obstime",
  n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE
)


UJM(
  formFixed = Y1 ~ obstime, formRandom = ~obstime,
  formGroup = ~id, formSurv = Surv(survtime, CR) ~ w1 + x1,
  dataLong, dataSurv, K = 15, model = "quadratic", Obstime = "obstime",
  n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1,
  DIC = TRUE, quiet = FALSE
)

}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
