% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS.R
\name{VS}
\alias{VS}
\title{Variable Selection Joint Modeling}
\usage{
VS(
  formFixed,
  formRandom,
  formGroup,
  formSurv,
  nmark,
  K1 = K1,
  K2 = K2,
  model = model,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  Obstime = "obstime",
  Method = "DS",
  ncl = ncl,
  DIC = TRUE,
  quiet = FALSE,
  dataLong,
  dataSurv
)
}
\arguments{
\item{formFixed}{a list of formulas for fixed part of longitudinal model}

\item{formRandom}{a list of formulas for random part of longitudinal model}

\item{formGroup}{a list of formulas specifying the cluster variable for Y (e.g. = list (~ subject, ~ subject,...))}

\item{formSurv}{formula for survival model}

\item{nmark}{the number of longitudinal markers}

\item{K1}{Number of nodes and weights for calculating Gaussian quadrature in the first stage.}

\item{K2}{Number of nodes and weights for calculating Gaussian quadrature in the second stage.}

\item{model}{a list of the models for the longitudinal part which includes "linear" or "quadratic".}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{Obstime}{the observed time in longitudinal data}

\item{Method}{the method for variable selection including "CS" for continues spike and "DS" for Dirac spike.}

\item{ncl}{the number of nodes to be forked for parallel computing}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=var(deviance) / 2 is used.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}
}
\value{
\itemize{
\item MCMC chains for the unknown parameters
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 25\% list of posterior median for each parameter
\item 50\% list of posterior median for each parameter
\item 75\% list of posterior median for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
}
}
\description{
Employ a two-stage approach to efficiently select variables for joint modeling of multiple longitudinal markers and competing risks outcomes within a Bayesian framework.
}
\details{
A two-stage variable selection approach within a Bayesian framework is designed to handle multiple longitudinal measurements and competing risks outcomes. In the first stage, we address the estimation of a one-marker joint model for the event along with each longitudinal marker. Leveraging these estimates, we derive predictions for the expected values or slopes of individual marker trajectories. In the second stage, we utilize a proportional hazard model that incorporates the expected current values and/or slopes of all markers as time-dependent covariates. Here, we adopt a spike-and-slab prior to effectively select crucial longitudinal markers and covariates, enhancing the model's predictive power and interpretability. All implementations have been done with JAGS.
}
\examples{
rm(list=ls())
data(dataLong)
data(dataSurv)
set.seed(2)
INDTRAIN <- sample(dataSurv$id, 0.5 * (dim(dataSurv)[1]))
INDVALID <- dataSurv$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong,
  dataLong$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv,
  dataSurv$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong,
  dataLong$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv,
  dataSurv$id \%in\% INDVALID
)



formFixed <- list(Y1 ~ obstime, Y2 ~ obstime, Y3 ~ obstime)
formRandom <- list(~obstime, ~obstime, ~obstime)
formGroup <- list(~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
model <- list("intercept", "linear", "quadratic")



VS <- VS(formFixed, formRandom, formGroup, formSurv,
  nmark = 3, K1 = 15, K2 = 15,
  model = model, n.chains = 1, n.iter = 100, n.burnin = 50,
  n.thin = 1, Obstime = "obstime", Method = "DS", ncl = 2,
  DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

SVS <- SVS(VS)

Step2 <- VS2(VS,
  Method = "LBFDR", n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1, dataLong = dataLong_t, dataSurv = dataSurv_t
)

DP <- DP(VS, Step2,
  Method = "LBFDR", s = 0.1, t = 0.5, n.chains = 1, n.iter = 100, n.burnin = 50,
  n.thin = 1,cause_main=1,
  DIC = TRUE, quiet = FALSE, dataLong = dataLong_v, dataSurv = dataSurv_v
)



######################################################################
\dontrun{
library(DPCri);library(survival)

Criteria(s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
         CR = dataSurv_v$CR, P = DP$DP[,2], cause = 1)



  rm(list=ls())
  set.seed(2)
  INDTRAIN=sample(dataSurv$id,.9*(dim(dataSurv)[1]))
  dataLong_t=subset(dataLong,
                    dataLong$id \%in\% INDTRAIN)

  dataSurv_t=subset(dataSurv,
                    dataSurv$id \%in\% INDTRAIN)
  names(dataSurv_t)
  formFixed <- list(
    Y1 ~ obstime + x1 + x2, Y2 ~ obstime + x1 + x2, Y3 ~ obstime + x1 + x2,
    Y4 ~ obstime + x1 + x2, Y5 ~ obstime + x1 + x2, Y6 ~ obstime + x1 + x2,
    Y7 ~ obstime + x1 + x2, Y8 ~ obstime + x1 + x2, Y9 ~ obstime + x1 + x2,
    Y10 ~ obstime + x1 + x2
  )
  formRandom <- list(
    ~obstime, ~obstime, ~obstime, ~obstime, ~obstime, ~obstime,
    ~obstime, ~obstime, ~obstime, ~obstime
  )
  formGroup <- list(~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id)
  formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
  model <- list(
    "linear", "linear", "linear", "linear", "linear", "linear",
    "linear", "linear", "quadratic", "quadratic"
  )



  VS_DS <- VS(formFixed, formRandom, formGroup, formSurv,
              nmark = 10, K1 = 15, K2 = 15,
              model = model, n.chains = 1, n.iter = 2000, n.burnin = 1000,
              n.thin = 1, Obstime = "obstime", Method = "DS",ncl=4,
              DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
  )

  SVS_DS <- SVS(VS_DS)



  VS_CS <- VS(formFixed, formRandom, formGroup, formSurv,
              nmark = 10, K1 = 15, K2 = 15,
              model = model, n.chains = 1, n.iter = 2000, n.burnin = 1000,
              n.thin = 1, Obstime = "obstime", Method = "CS",
              DIC = TRUE, quiet = FALSE, dataLong, dataSurv
  )

  SVS_CS <- SVS(VS_CS)

}




}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
