% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS2.R
\name{VS2}
\alias{VS2}
\title{Variable Selection Joint Modeling 2}
\usage{
VS2(
  object,
  Method = "LBFDR",
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  n.thin = max(1, floor((n.iter - n.burnin)/1000)),
  DIC = TRUE,
  quiet = FALSE,
  dataLong,
  dataSurv
)
}
\arguments{
\item{object}{an object inheriting from class VSJM}

\item{Method}{the method for variable selection including "LBFDR" for LBFDR and "BF" for Bayes factor.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=var(deviance) / 2 is used.}

\item{quiet}{Logical, whether to suppress stdout in jags.model().}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}
}
\value{
\itemize{
\item MCMC chains for the unknown parameters
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 25\% list of posterior median for each parameter
\item 50\% list of posterior median for each parameter
\item 75\% list of posterior median for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
}
}
\description{
Running a time-dependent proportional hazard model on the results of VS
}
\details{
After variable selection by VS, the second stage is re-estimated by replacing CS or Ds with non-informative normal distributions as priors for the association parameters of the selected markers and the regression coefficients of the selected covariates.
}
\examples{
rm(list=ls())
data(dataLong)
data(dataSurv)
set.seed(2)
INDTRAIN <- sample(dataSurv$id, 0.5 * (dim(dataSurv)[1]))
INDVALID <- dataSurv$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong,
  dataLong$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv,
  dataSurv$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong,
  dataLong$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv,
  dataSurv$id \%in\% INDVALID
)



formFixed <- list(Y1 ~ obstime, Y2 ~ obstime, Y3 ~ obstime)
formRandom <- list(~obstime, ~obstime, ~obstime)
formGroup <- list(~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
model <- list("intercept", "linear", "quadratic")



VS <- VS(formFixed, formRandom, formGroup, formSurv,
  nmark = 3, K1 = 15, K2 = 15,
  model = model, n.chains = 1, n.iter = 100, n.burnin = 50,
  n.thin = 1, Obstime = "obstime", Method = "DS", ncl = 2,
  DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

SVS <- SVS(VS)

Step2 <- VS2(VS,
  Method = "LBFDR", n.chains = 2, n.iter = 100, n.burnin = 50,
  n.thin = 1, dataLong = dataLong_t, dataSurv = dataSurv_t
)

DP <- DP(VS, Step2,
  Method = "LBFDR", s = 0.1, t = 0.5, n.chains = 1, n.iter = 100, n.burnin = 50,
  n.thin = 1,cause_main=1,
  DIC = TRUE, quiet = FALSE, dataLong = dataLong_v, dataSurv = dataSurv_v
)



######################################################################
\dontrun{
library(DPCri);library(survival)

Criteria(s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
         CR = dataSurv_v$CR, P = DP$DP[,2], cause = 1)

######################################################################
rm(list=ls())
data(dataLong_m)
data(dataSurv_m)
set.seed(2)
INDTRAIN <- sample(dataSurv_m$id, 0.5 * (dim(dataSurv_m)[1]))
INDVALID <- dataSurv_m$id[-INDTRAIN]
dataLong_t <- subset(
  dataLong_m,
  dataLong_m$id \%in\% INDTRAIN
)
dataSurv_t <- subset(
  dataSurv_m,
  dataSurv_m$id \%in\% INDTRAIN
)
names(dataSurv_t)

dataLong_v <- subset(
  dataLong_m,
  dataLong_m$id \%in\% INDVALID
)
dataSurv_v <- subset(
  dataSurv_m,
  dataSurv_m$id \%in\% INDVALID
)



formFixed <- list(Y1 ~ obstime, Y2 ~ obstime, Y3 ~ obstime)
formRandom <- list(~obstime, ~obstime, ~obstime)
formGroup <- list(~id, ~id, ~id)
formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
model <- list("intercept", "linear", "quadratic")



VS <- VS(formFixed, formRandom, formGroup, formSurv,
         nmark = 3, K1 = 15, K2 = 15,
         model = model, n.chains = 1, n.iter = 50, n.burnin = 20,
         n.thin = 1, Obstime = "obstime", Method = "DS", ncl = 2,
         DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
)

SVS <- SVS(VS)

Step2 <- VS2(VS,
             Method = "LBFDR", n.chains = 2, n.iter = 100, n.burnin = 50,
             n.thin = 1, dataLong = dataLong_t, dataSurv = dataSurv_t
)

DP <- DP(VS, Step2,
         Method = "LBFDR", s = 0.1, t = 0.5, n.chains = 1, n.iter = 100, n.burnin = 50,
         n.thin = 1,cause_main=1,
         DIC = TRUE, quiet = FALSE, dataLong = dataLong_v, dataSurv = dataSurv_v
)


Criteria(s = 0.1, t = 0.5, Survt = dataSurv_v$survtime,
         CR = dataSurv_v$CR, P = DP$DP[,2], cause = 1)

######################################################################

  rm(list=ls())
  set.seed(2)
  INDTRAIN=sample(dataSurv$id,.9*(dim(dataSurv)[1]))
  dataLong_t=subset(dataLong,
                    dataLong$id \%in\% INDTRAIN)

  dataSurv_t=subset(dataSurv,
                    dataSurv$id \%in\% INDTRAIN)
  names(dataSurv_t)
  formFixed <- list(
    Y1 ~ obstime + x1 + x2, Y2 ~ obstime + x1 + x2, Y3 ~ obstime + x1 + x2,
    Y4 ~ obstime + x1 + x2, Y5 ~ obstime + x1 + x2, Y6 ~ obstime + x1 + x2,
    Y7 ~ obstime + x1 + x2, Y8 ~ obstime + x1 + x2, Y9 ~ obstime + x1 + x2,
    Y10 ~ obstime + x1 + x2
  )
  formRandom <- list(
    ~obstime, ~obstime, ~obstime, ~obstime, ~obstime, ~obstime,
    ~obstime, ~obstime, ~obstime, ~obstime
  )
  formGroup <- list(~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id, ~id)
  formSurv <- survival::Surv(survtime, CR) ~ w1 + x1
  model <- list(
    "linear", "linear", "linear", "linear", "linear", "linear",
    "linear", "linear", "quadratic", "quadratic"
  )



  VS_DS <- VS(formFixed, formRandom, formGroup, formSurv,
              nmark = 10, K1 = 15, K2 = 15,
              model = model, n.chains = 1, n.iter = 2000, n.burnin = 1000,
              n.thin = 1, Obstime = "obstime", Method = "DS",ncl=4,
              DIC = TRUE, quiet = FALSE, dataLong_t, dataSurv_t
  )

  SVS_DS <- SVS(VS_DS)



  VS_CS <- VS(formFixed, formRandom, formGroup, formSurv,
              nmark = 10, K1 = 15, K2 = 15,
              model = model, n.chains = 1, n.iter = 2000, n.burnin = 1000,
              n.thin = 1, Obstime = "obstime", Method = "CS",
              DIC = TRUE, quiet = FALSE, dataLong, dataSurv
  )

  SVS_CS <- SVS(VS_CS)

}




}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
